feat: rich error message POC
