+++ projects/playground/src/app/lazy/lazy.module.ts
@@ -1,7 +1,7 @@
 import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
 import { ReactiveFormsModule } from '@angular/forms';
-import { NgxErrorsModule } from '@ngspot/ngx-errors';
+import { NgxErrorsModule } from '../../../../ngx-errors/src/public-api';
 import { LazyRoutingModule } from './lazy-routing.module';
 import { LazyComponent } from './lazy.component';
